:hammer: refactor: 디렉토리 모듈 정리
