+++ .gitignore
@@ -1,3 +1,4 @@
 config.json
 *.log
 __pycache__
+test
