--- crawling/ex_links
@@ -1,2 +1,0 @@
-https://n.news.naver.com/article/025/0003340254?ntype=RANKING
-https://n.news.naver.com/article/277/0005377731?cds=news_media_pc
