+++ crawling/utils/config.py
@@ -1,11 +1,13 @@
 import json
 import logging
 import logging.config
+import sys
+sys.path.append('../')
 
 # 네이버 api 사용자 정보 가져오는 함수
 def getConfigData(data : str) -> map :
 	try:
-		with open("./config/config.json", 'r') as f :
+		with open("../config/config.json", 'r') as f :
 			jsonData = json.load(f)
 	except Exception as e :
 		logging.getLogger('__main__').error(e)
